% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_footprinting.R
\name{atac_motif_dist}
\alias{atac_motif_dist}
\title{Calculate distances of Tn5 insertions from motif.}
\usage{
atac_motif_dist(tn_signal_gr, motif_pos_gr, range = 200)
}
\arguments{
\item{tn_signal_gr}{A GRanges object with tn5 insertions. See the read_atac_bam_tn function.}

\item{motif_pos_gr}{A GRanges object with the positions of the PWM in ATAC-seq peaks.}

\item{range}{The distance in bases flanking the the motif to calculate signal.}
}
\value{
A numeric vector of distances of Tn5 insertion sites to motif centre.
}
\description{
Calculate distances of Tn5 insertions from motif.
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
ctcf <- read.table("data/ctcf.pwm") \%>\% as.matrix()
peaks <- read_bed("data/atac_peaks.bed")
mpos <- motif_gr(gr, ctcf, Mmusculus)
tn <- read_atac_bam_tn(bam_file = "data/atac.bam", yieldSize = 1e6, which=peaks)
dists <- atac_motif_dist(tn, mpos)
head(dists)
}
}

