% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_centres_bigwig.R
\name{bam_to_centres_bw}
\alias{bam_to_centres_bw}
\title{Write a bigwig file of Tn5 insertion depth}
\usage{
bam_to_centres_bw(bam_file, file, max_insert = 2000, nuc_range = c(189,
  247), scale_cpm = FALSE, ...)
}
\arguments{
\item{bam_file}{Path to an ATAC-seq paired-end BAM file.}

\item{file}{Path for output bigwig file. Extension should be .bw or .bigwig.}

\item{max_insert}{The Maximum insert size for read fragments.}

\item{nuc_range}{The nucleosome spaning fragment size range.
#' @param scale_cpm Logical. Should the output be scaled to number of 
million mapped reads?}
}
\value{
No object is returned. A file is written.
}
\description{
Write a bigwig file of Tn5 insertion depth
}

