% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_insertions_bigwig.R
\name{bam_to_insertions_bw}
\alias{bam_to_insertions_bw}
\title{Write a bigwig file of Tn5 insertion depth}
\usage{
bam_to_insertions_bw(bam_file, file, yieldSize = 1e+06, scale_cpm = FALSE,
  ...)
}
\arguments{
\item{bam_file}{Path to an ATAC-seq paired-end BAM file.}

\item{file}{Path for output bigwig file. Extension should be .bw or .bigwig.}

\item{yieldSize}{Number of records to yield each time the file is read from with scanBam.
See ?BamFile}

\item{scale_cpm}{Logical. Should the output be scaled to number of 
million mapped reads?}
}
\value{
No object is returned. A file is written.
}
\description{
Write a bigwig file of Tn5 insertion depth
}

