% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dinuc_freq.R
\name{calc_dinuc_freq}
\alias{calc_dinuc_freq}
\title{Calculate the dinucleotide insertion frequency for Tn5 insertions.}
\usage{
calc_dinuc_freq(ins_gr, genome)
}
\arguments{
\item{ins_gr}{A GRanges object. 
Typically of ATAC-seq Tn5 insertions generated with the function read_atac_insertions.}

\item{genome}{A BSgenome object.}
}
\value{
A data.frame containing dinucleotide counts and precentages.
}
\description{
Calculate the dinucleotide insertion frequency for Tn5 insertions.
}

