% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_frip.R
\name{calc_frip}
\alias{calc_frip}
\title{Calculate the fraction of reads in peaks.}
\usage{
calc_frip(signal, peaks)
}
\arguments{
\item{signal}{A GRanges object. Typically of ATAC-seq Tn5 insertions.}

\item{peaks}{A GRanges object. Typically of ATAC-seq peaks.}
}
\value{
Fraction of ranges in signal overlapping peaks. Numeric.
}
\description{
Calculate the fraction of reads in peaks.
}

