% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ins_bias.R
\name{calc_ins_bias}
\alias{calc_ins_bias}
\title{Calculate the PWM for Tn5 insertions.}
\usage{
calc_ins_bias(ins_pwm, regions_gr, genome)
}
\arguments{
\item{ins_pwm}{A positional weight matrix generated with the function calc_ins_pwm.
Must be a matrix of 21 columns and 4 rows.}

\item{regions_gr}{A GRanges object of regions to calculate insertion bias scores.}

\item{genome}{A BSgenome object.
site for calculating the PWM.}
}
\value{
A vector of per-base bias scores for each range in regions_gr.
}
\description{
Calculate the PWM for Tn5 insertions.
}

