% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_motif_dist.R
\name{calc_motif_dist}
\alias{calc_motif_dist}
\title{Calculate distances of Tn5 insertions from motif.}
\usage{
calc_motif_dist(signal_gr, motif_pos_gr, range = 200)
}
\arguments{
\item{signal_gr}{A GRanges object with tn5 insertions. See the read_atac_bam_tn function.}

\item{motif_pos_gr}{A GRanges object with the positions of the PWM in ATAC-seq peaks.}

\item{range}{The distance in bases flanking the the motif to calculate signal.}
}
\value{
A numeric vector of distances of Tn5 insertion sites to motif centre.
}
\description{
Calculate distances of Tn5 insertions from motif.
}

