% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_v.R
\name{calc_v}
\alias{calc_v}
\title{Calculate distances of Tn5 insertions from motif.}
\usage{
calc_v(frags_gr, motif_pos_gr, flank = 250, max_frag = 300)
}
\arguments{
\item{frags_gr}{A GRanges object with ATAC-seq fragments.
See the read_atac_frags function.}

\item{motif_pos_gr}{A GRanges object with the positions of the PWM in ATAC-seq peaks.
See the motif_gr function.}

\item{flank}{Numeric of length 1. Number of bases flanking motif to plot.}

\item{max_frag}{Integer of length 1. The maximum fragment size to include in plot.}
}
\value{
A data.frame with 2 columns.
frag_size column contains the ATAC-seq fragment size.
The distance column contains the distance from the fragment centure to the centre of
the closest motif in motif_pos_gr.
}
\description{
Calculate distances of Tn5 insertions from motif.
}

