% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{gr_centre_motif}
\alias{gr_centre_motif}
\title{Centre ranges on motif}
\usage{
gr_centre_motif(gr, pwm, bs_genome, min.score = "85\%")
}
\arguments{
\item{gr}{A GRanges object}

\item{pwm}{A positional weight matrix. Should be a Matrix object with 4 rows.}

\item{bs_genome}{A BSgenome object.}

\item{min.score}{Character. Minimum alignment score percentage for the PWM.}
}
\value{
A GRanges object. Ranges are of width=2, 
which are the centre of the PWM alignment.
}
\description{
Centre ranges on motif
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
gr_centre_motif(gr, ctcf.pwm, Mmusculus)
}
}

