% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_footprinting.R
\name{motif_gr}
\alias{motif_gr}
\title{Get the positions of the PWM in ATAC-seq peaks.}
\usage{
motif_gr(gr, pwm, genome = Mmusculus, min.score = "85\%")
}
\arguments{
\item{gr}{A GRanges object. Usually ranges of ATAC-seq peaks.}

\item{pwm}{A positional weight matrix of class Matirx.}

\item{genome}{A BSgenome object.}

\item{min.score}{Minimum alignment score for pwm. Default = "85%".}
}
\value{
A GRanges object with positions of PWM match in peaks.
}
\description{
Get the positions of the PWM in ATAC-seq peaks.
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
ctcf <- read.table("/data/ctcf.pwm") \%>\% as.matrix()
peaks <- read_bed("/data/atac_peaks.bed")
mpos <- motif_gr(peaks, ctcf, Mmusculus)
}
}

