% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_insert_size.R
\name{plot_insert_size}
\alias{plot_insert_size}
\title{Plot insert size historgram}
\usage{
plot_insert_size(frag_gr, ...)
}
\arguments{
\item{frag_gr}{A GRanges object of ATAC-seq fragments}
}
\value{
A plot to the current graphics device
}
\description{
Plot insert size historgram
}

