% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_summary.R
\name{range_summary}
\alias{range_summary}
\title{Get the signal from bigwig file for defined regions.}
\usage{
range_summary(bigwig, gr, range = 100, log = FALSE, aggregate = FALSE)
}
\arguments{
\item{bigwig}{Path to a bigwig file. Also accpets http paths.}

\item{gr}{A genomic ranges object of regions to obtain signal. All ranges should be the same width.}

\item{range}{The distance from the centre of gr to flank. Default=100.}

\item{log}{Logical. Is the signal in the bigwig file in log space? Default=FALSE.
If log = TRUE, the scores in the bigwig file are transformed by e^scores.}

\item{aggregate}{Logical. Should data for each position be added to get an aggregate score?}
}
\value{
A numeric vector of aggregate signal if aggregate = TRUE, which is the default. 
If aggregate = FALSE a data.frame of signal of all regions is returned.
}
\description{
Get the signal from bigwig file for defined regions.
}

