% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atac_insertions.R
\name{read_atac_insertions}
\alias{read_atac_insertions}
\title{Read BAM file Tn5 insertion positions into GRanges object}
\usage{
read_atac_insertions(bam_file, yieldSize = 1e+06, ...)
}
\arguments{
\item{bam_file}{Path to a BAM formatted file.}

\item{yieldSize}{The number of records from the BAM file to read in each chunk. See ?BamFile.}

\item{...}{Arguments passed to ScanBamParam.}
}
\value{
A GRanges object of Tn5 insertion positions
}
\description{
Read BAM file Tn5 insertion positions into GRanges object
}

