% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_gr_bed}
\alias{write_gr_bed}
\title{Write ranges in a GenomicRanges object to a BED file}
\usage{
write_gr_bed(gr, file, score = "0", name = ".")
}
\arguments{
\item{gr}{A GRanges object.}

\item{file}{Path for output bed file.}

\item{score}{An vector of scores for bed file.}

\item{name}{An optional vector of names for bed file.}
}
\description{
Write ranges in a GenomicRanges object to a BED file
}

