% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_insertions_bw}
\alias{write_insertions_bw}
\title{Write a bigwig file of Tn5 insertion depth}
\usage{
write_insertions_bw(gr, file, scale_cpm = FALSE)
}
\arguments{
\item{gr}{A GRanges object containing nucleosome positions.}

\item{file}{Path for output bigwig file. Extension should be .bw or .bigwig.}

\item{scale_cpm}{Logical. Should the output be scaled to number of 
nucleosome spanning reads.}
}
\description{
Write a bigwig file of Tn5 insertion depth
}

