% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ins_pwm.R
\name{calc_ins_pwm}
\alias{calc_ins_pwm}
\title{Calculate the PWM for Tn5 insertions.}
\usage{
calc_ins_pwm(ins_gr, genome, ...)
}
\arguments{
\item{ins_gr}{A GRanges object. 
Typically of ATAC-seq Tn5 insertions generated with the function read_atac_insertions.}

\item{genome}{A BSgenome object.}

\item{...}{Parameters for the PWM function in the Biostrings package.}
}
\value{
A PWM matirix containing dinucleotide counts and precentages.
}
\description{
Calculate the PWM for Tn5 insertions.
}

