% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_footprinting.R
\name{plot_scaled_footprint}
\alias{plot_scaled_footprint}
\title{Plot scaled ATAC-seq footprint signal.}
\usage{
plot_scaled_footprint(tn_dist, window = 200, ...)
}
\arguments{
\item{tn_dist}{A vector of distances of tn5 insertions relative to motif centre.
See atac_motif_dist function.}

\item{window}{The plotting window in bases.}

\item{...}{Optional arguments passed to plot function.}
}
\value{
An ATAC-seq footprint plot in the current graphics device. No values are returned.
}
\description{
Plot scaled ATAC-seq footprint signal.
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
ctcf <- read.table("data/ctcf.pwm") \%>\% as.matrix()
peaks <- read_bed("data/atac_peaks.bed")
mpos <- motif_gr(gr, ctcf, Mmusculus)
tn <- read_atac_bam_tn(bam_file = "data/atac.bam", yieldSize = 1e6, which=peaks)
dists <- atac_motif_dist(tn, mpos)
plot_scaled_footprint(dists)
}
}

