% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_v.R
\name{plot_v}
\alias{plot_v}
\title{Calculate distances of Tn5 insertions from motif.}
\usage{
plot_v(df, size = 1, alpha = 0.01, ...)
}
\arguments{
\item{df}{A data.frame returned from funciton calc_v.}

\item{size}{Numeric. Point size.}

\item{alpha}{Numeric value between 0 and 1. Point transparency.}

\item{...}{Arguments for geom_point in ggplot.}
}
\value{
An object of class 'ggplot'.
}
\description{
Calculate distances of Tn5 insertions from motif.
}

