% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atac_frags.R
\name{read_atac_frags}
\alias{read_atac_frags}
\title{Read BAM file aligned fragments into GRanges object}
\usage{
read_atac_frags(bam_file, max_insert = 2000, yieldSize = 1e+06, ...)
}
\arguments{
\item{bam_file}{Path to a BAM formatted file.}

\item{max_insert}{Maximum insert size allowed. Default is reccomended.}

\item{yieldSize}{The number of reads retreived from the BAM file in each chunk.
See ?BamFile in Rsamtools.}

\item{...}{Arguments passed to ScanBamParam.}
}
\value{
A GRanges object of aligned fragments
}
\description{
Read BAM file aligned fragments into GRanges object
}

