% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_atac_pos}
\alias{read_atac_pos}
\title{Read BAM file Tn5 insertion positions into GRanges object}
\usage{
read_atac_pos(bam_file, yieldSize = 1e+05, ...)
}
\arguments{
\item{bam_file}{Path to a BAM formatted file.}

\item{regions_gr}{A GenomicRanges object of genomic regions of data to be retreived.}
}
\value{
A GRanges object of Tn5 insertion positions
}
\description{
Read BAM file Tn5 insertion positions into GRanges object
}
\examples{
\dontrun{
tn <- import_atac_pos("test_100k.bam")
}
}

