% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.R
\name{read_bed}
\alias{read_bed}
\title{Read BED formatted file to GRanges object}
\usage{
read_bed(bed_file)
}
\arguments{
\item{bed_file}{Path Path to BED formatted file.}
}
\value{
A GRanges object
}
\description{
Read BED formatted file to GRanges object
}

