% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_gr_bed}
\alias{write_gr_bed}
\title{Write ranges in a GenomicRanges object to a BED file}
\usage{
write_gr_bed(gr, file)
}
\arguments{
\item{gr}{A GRanges object.}

\item{file}{Path for output bed file.}
}
\description{
Write ranges in a GenomicRanges object to a BED file
}

